int->unsigned long int

+++ 数据结构/结构体/简易通讯录.h
@@ -4,7 +4,7 @@
 
 struct student
 {
-    int id;
+    unsigned long int id;
     struct bir
     {
         int year;
